refactor: refactor feature detail
